+++ px-checkout/src/main/java/com/mercadopago/android/px/core/internal/OnPaymentListener.kt
@@ -1,0 +1,5 @@
+package com.mercadopago.android.px.core.internal
+
+import com.mercadopago.android.px.core.SplitPaymentProcessor
+
+internal interface OnPaymentListener : SplitPaymentProcessor.OnPaymentListener