+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/viewmodel/DebitCardDescriptorModel.java
@@ -61,7 +61,7 @@ public final class DebitCardDescriptorModel extends PaymentMethodDescriptorView.
             .toSpannable();
 
         new AmountLabeledFormatter(spannableStringBuilder, context)
-            .withTextColor(ContextCompat.getColor(context, R.color.ui_meli_black))
+            .withTextColor(ContextCompat.getColor(context, R.color.px_expressCheckoutTextColor))
             .withSemiBoldStyle()
             .apply(amount);
     }