+++ px-checkout/src/main/java/com/mercadopago/android/px/tracking/internal/model/InitData.java
@@ -5,6 +5,7 @@ import androidx.annotation.Nullable;
 import com.mercadopago.android.px.configuration.AdvancedConfiguration;
 import com.mercadopago.android.px.configuration.PaymentConfiguration;
 import com.mercadopago.android.px.internal.repository.PaymentSettingRepository;
+import com.mercadopago.android.px.internal.util.PaymentConfigurationUtil;
 import com.mercadopago.android.px.preferences.CheckoutPreference;
 
 @SuppressWarnings("unused")
@@ -23,7 +24,8 @@ public final class InitData extends TrackingMapModel {
         checkoutPreferenceId = preferenceId;
         checkoutPreference = preference;
         expressEnabled = advancedConfiguration.isExpressPaymentEnabled();
-        splitEnabled = preference != null && paymentConfiguration.getPaymentProcessor().supportsSplitPayment(preference);
+        splitEnabled = preference != null &&
+            PaymentConfigurationUtil.getPaymentProcessor(paymentConfiguration).supportsSplitPayment(preference);
         escEnabled = true;
     }