+++ px-services/src/main/java/com/mercadopago/android/px/model/internal/PaymentTypeChargeRuleDM.kt
@@ -5,4 +5,6 @@ import java.math.BigDecimal
 data class PaymentTypeChargeRuleDM (
     val paymentTypeId: String,
     val charge: BigDecimal,
-    val message: String?)
+    val message: String?,
+    val taxable: Boolean
+)